% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o3_max_mm8h_d.R
\docType{data}
\name{o3_max_mm8h_d}
\alias{o3_max_mm8h_d}
\title{Tibble con i dati di o3_max_mm8h_d per 15 stazioni della regione PIEMONTE}
\format{
Un tibble con 8 colonne e 10965 osservazioni
}
\usage{
o3_max_mm8h_d
}
\description{
Tibble con i dati di o3_max_mm8h_d per 15 stazioni della regione PIEMONTE
}
\keyword{datasets}
